requirements.txt
# data science
numpy
pandas

# img

image
opencv-python

# Uncomment if you use sklearn
# scikit-learn

# Install the correct TensorFlow version
# Uncomment if you use TensorFlow
#tensorflow==2.15.0   # For both Intel and Apple silicon

# Google Cloud
google-cloud-storage

python-dotenv

# API
fastapi
uvicorn

pillow

shap
matplotlib

Response

os

